No need for lib directory.
The index.js file has no code, there is no argument which might defend its existence.
